Reworked visuals
- Visuals no longer render themselves; they just state their intended appearance
- GameSprite removed/merged into visual
- Visuals don't _need_ to be animated - there's AnimationVisual for that
- GameBatcher no longer has OptimisedDrawers
- There is a CustomRenderer visual that replaced the need to a OptimisedDrawer
- Some visuals can have sub visuals attached with GroupVisual

Added a font file for Font1 that lets it be used as a BitmapFont. I don't really know how BitmapFonts work yet so it's probably wrong. When characters are hit, they display the damage done.

As a side change, ColorMaths has been updated with some more functionality. Tests were written to ensure that it works correctly (it was broken before)