Small improvements to SimpleFog
The fog starts out being transparent. This is because when starting from a save, the fog doesn't get updated until the hero moves (which is a bug that needs fixing) and so I can't tell what is happening.

The fog also has a much more efficient method of loading the altered pixmap