Added implementation of items
Items are now added. In addition to items, there is ItemStacks which represent a quantity of items.

The backpack has been added - it is currently only an in game thing as I am yet to make the renderer.

The behaviour of item visuals is rather poor. Currently, Item implements the WorldVisual interface. However, items do not match the implementation of visuals very well since they have no position. Thus, a copy of the visual has to be made in the heap (which currently also doesn't know it's position :P so the Level render has to manually update it's position (but I think I will make the heap know in the future so that wil fix that bit of yick)).