Improved UI
Moved all assets into sub folders because they were to hard to find

Got the standard pixel dungeon ui skin working and it is being used on the backpack window.

Added a "TabbedWindow". It's a game window with tabs :)

GameWindow now fits the the content instead of filling the entire screen