Overhauled Items
Items are no longer immutable. There are many item interfaces that specify the different behaviours of an item which replaced the need for ItemStacks. This also creates a convenient method of specifying  what things are sellable/stackable/upgradeable without cluttering every item with annoying isSellabe()/isUnique().

Heaps are now actors. This means the do not require any custom drawing code in the LevelRenderer. Heaps also have a HeapType which specifies how the heap looks. In a future release, it will also specify how the item is picked up (e.g. if it needs a key to be opened).

I'm in the middle of overhauling the Hero's backpack. The BackPack.java has been renamed to SomeSlots.java (still need a better name). Then Hero's will have an inventory which contains 2 SomeSlots for the equipped slots, the backpack slots.