Improved logic in LevelUpdater
The level updater no longer requires knowing if actors are user controlled. It also no longer tells an action to update twice when the action finishes on the first tick.

Because of these changes, Actor can have the isUserControlled() method removed, and ChangeGoalAction can actually check if it is being used twice in a row.

I've increased the time to take a step just a little: it seems to make it a bit smoother.