Mob Spawing!
Added a RandomCategory. It simply returns a random value given to it. Using that and a MobConstructor interface, a Bestiary is formed.

Every Level now has a bestiary and every LevelDescriptor also has a Bestiary which is passed onto the Level.

Mobs are currently spawned at time 0 even though everyone else is already at some large time. Thus, they get lots of turns when they spawn. I'll fix this by switching over to an energy system.